finished generateSentimentScores

+++ data/generateSentimentScores.py
@@ -9,44 +9,66 @@ Data used from New York Times:
 
 import csv
 import json
+import logging
 import nltk
 import os
 import shutil
-from src.finbert import predict
+import sys
+import time
+from tqdm import tqdm
 from transformers import AutoModelForSequenceClassification
+
+sys.path.insert(1, "../src")
+from finbert import predict
+
+
+def useFinBert(article):
+    for text in [article["abstract"], article["headline"]["main"]]:
+        if text is not None and text != "" and not str.isspace(text):
+            for column in predict(text, FinBert).iterrows():
+                dataResults.append([row[0], column[1]["sentence"], column[1]["prediction"], column[1]["sentiment_score"]])
 
 
 if __name__ == "__main__":
 
     # Parameter Declaration
     datasetName = "kdd17"  # Can be either "kdd17" or "stocknet"
+    fromStart = False
 
     # Variable Declaration
-    stockCodes = os.listdir(datasetName + "/NYT-Business")
+    stockCodes = os.listdir(datasetName + "/NYT-Business/ourpped")
     FinBert = AutoModelForSequenceClassification.from_pretrained("../models/FinBert", cache_dir=None, num_labels=3)
 
+    # nltk.download("punkt")
+    logging.disable()
+    if fromStart:
+        if os.path.exists(datasetName + "/SentimentScores/NYT-Business"):
+            shutil.rmtree(datasetName + "/SentimentScores/NYT-Business")
 
-    nltk.download("punkt")
+        os.mkdir(datasetName + "/SentimentScores/NYT-Business")
 
-    if os.path.exists(datasetName + "/SentimentScores/NYT-Business"):
-        shutil.rmtree(datasetName + "/SentimentScores/NYT-Business")
+    completed = os.listdir(datasetName + "/SentimentScores/NYT-Business")
 
-    os.mkdir(datasetName + "/SentimentScores/NYT-Business")
+    for stock in tqdm(os.listdir(datasetName + "/NYT-Business/ourpped")):
+        if stock in completed:
+            continue
+        start = time.process_time()
+        dataResults = [["Date", "Sentence", "Prediction", "sentimentScore"]]
+        with open(datasetName + "/NYT-Business/ourpped/" + stock, encoding="utf-8") as file:
+            csvReader = csv.reader(file)
+            next(csvReader)  # Skip headers
 
-    for stock in stockCodes:
-        os.mkdir(datasetName + "/SentimentScores/NYT-Business/" + stock)
-        for date in os.listdir(datasetName + "/NYT-Business/" + stock):
-            dateResults = [["Sentence", "Prediction", "sentimentScore"]]
-            with open(datasetName + "/NYT-Business/" + stock + '/' + date, encoding="utf-8") as file:
-                data = json.load(file)
-                completed = []
-                for text in [data["abstract"], data["snippet"], data["lead_paragraph"], data["headline"]["main"]]:
-                    if text is not None and text != "" and not str.isspace(text) and text not in completed:
-                        completed.append(text)
-                        temp = predict(text, FinBert)
-                        for column in temp.iterrows():
-                            dateResults.append([column[1]["sentence"], column[1]["prediction"], column[1]["sentiment_score"]])
-            with open(datasetName + "/SentimentScores/NYT-Business/" + stock + '/' + date[:-5] + ".csv", "w+", encoding="utf-8", newline='') as file:
-                csvWriter = csv.writer(file)
-                csvWriter.writerows(dateResults)
-            exit()
+            for row in tqdm(csvReader):
+
+                data = json.loads(row[1])
+                if type(data) is list:
+                    for article in data:
+                        useFinBert(article)
+                else:
+                    useFinBert(data)
+
+        with open(datasetName + "/SentimentScores/NYT-Business/" + stock, "w+", encoding="utf-8", newline='') as file:
+            csvWriter = csv.writer(file)
+            csvWriter.writerows(dataResults)
+        print(time.process_time() - start)
+        exit()
